画面遷移ほぼ完成
画面遷移できるように各ページを修正しました。

+++ WEB-INF/src/food_management/ShokuzaiBeanMapping.java
@@ -12,13 +12,13 @@ public class ShokuzaiBeanMapping implements ResultSetBeanMapping<ShokuzaiBean> {
 		sbean.setPk_id(rs.getInt("pk_id"));
 		sbean.setMaterials_name(rs.getString("materials_name"));
 		sbean.setItem_name(rs.getString("Item_name"));
-		sbean.setQuantity(rs.getInt("quantity"));
+		sbean.setQuantity(rs.getDouble("quantity"));
 		sbean.setOpen_check(rs.getInt("open_check"));
 		sbean.setSave_method(rs.getInt("save_method"));
 		sbean.setGenre_id(rs.getInt("genre_id"));
 		sbean.setFilling_date(rs.getString("filling_date"));
 		sbean.setExpiration_date(rs.getString("expiration_date"));
-		sbean.setVolume(rs.getInt("volume"));
+		sbean.setVolume(rs.getDouble("volume"));
 		sbean.setUnit_id(rs.getInt("unit_id"));
 		sbean.setDefault_volume(rs.getDouble("default_volume"));
 		sbean.setShokuzai_favorite(rs.getInt("shokuzai_favorite"));
